Branch riccardo

+++ shelfy/app/src/main/java/com/example/shelfy/data/remote/responses/VolumeInfo.kt
@@ -3,7 +3,7 @@ package com.example.shelfy.data.remote.responses
 data class VolumeInfo(
     val allowAnonLogging: Boolean,
     val authors: List<String>,
-    val averageRating: Int,
+    val averageRating: Float,
     val canonicalVolumeLink: String,
     val categories: List<String>,
     val contentVersion: String,
